# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/categories": {
      "get": {
        "operationId": "findCategories",
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Category"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "categories"
        ]
      },
      "post": {
        "operationId": "createCategory",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Category"
                }
              }
            }
          },
          "400": {
            "description": ""
          }
        },
        "tags": [
          "categories"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/users": {
      "get": {
        "operationId": "findUsers",
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  }
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/users/{userId}": {
      "get": {
        "operationId": "findUser",
        "summary": "",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ]
      },
      "patch": {
        "operationId": "updateUser",
        "summary": "",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteUser",
        "summary": "",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": ""
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/products": {
      "get": {
        "operationId": "getProducts",
        "summary": "",
        "parameters": [
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No products found"
          }
        },
        "tags": [
          "products"
        ]
      },
      "post": {
        "operationId": "createProduct",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/products/{productId}": {
      "get": {
        "operationId": "getProductById",
        "summary": "",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "tags": [
          "products"
        ]
      },
      "patch": {
        "operationId": "updateProduct",
        "summary": "",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteProduct",
        "summary": "",
        "parameters": [
          {
            "name": "productId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "products"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/records": {
      "get": {
        "operationId": "getRecords",
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Record"
                  }
                }
              }
            }
          },
          "404": {
            "description": "No products found"
          }
        },
        "tags": [
          "records"
        ]
      },
      "post": {
        "operationId": "createRecord",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRecordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          }
        },
        "tags": [
          "records"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/records/{recordId}": {
      "get": {
        "operationId": "getRecordById",
        "summary": "",
        "parameters": [
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          },
          "404": {
            "description": "Record not found"
          }
        },
        "tags": [
          "records"
        ]
      },
      "patch": {
        "operationId": "updateRecord",
        "summary": "",
        "parameters": [
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRecordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Record"
                }
              }
            }
          }
        },
        "tags": [
          "records"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteRecord",
        "summary": "",
        "parameters": [
          {
            "name": "recordId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "tags": [
          "records"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/files/upload": {
      "post": {
        "operationId": "uploadFile",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "register",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "login",
        "summary": "",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "findMe",
        "summary": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindMeResponse"
                }
              }
            }
          }
        },
        "tags": [
          "auth"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Veggies Api",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "Category": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "categoryId",
          "name",
          "createdAt"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 10
          }
        },
        "required": [
          "name"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "taxId": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "name",
          "email",
          "address",
          "phone",
          "taxId",
          "isAdmin",
          "createdAt"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 30
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "minLength": 6
          },
          "taxId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 8
          }
        }
      },
      "Product": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number"
          },
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "name": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "records": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "latestPrice": {
            "type": "number"
          },
          "seller": {
            "$ref": "#/components/schemas/User"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "productId",
          "category",
          "name",
          "image",
          "records",
          "seller",
          "createdAt"
        ]
      },
      "InitialRecordDto": {
        "type": "object",
        "properties": {
          "unitOfMeasure": {
            "type": "string"
          },
          "unitPrice": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "unitOfMeasure",
          "unitPrice"
        ]
      },
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "minimum": 1
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "image": {
            "type": "string"
          },
          "categoryId": {
            "type": "number",
            "minimum": 1
          },
          "records": {
            "default": [],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InitialRecordDto"
            }
          }
        },
        "required": [
          "userId",
          "name",
          "image",
          "categoryId"
        ]
      },
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "categoryId": {
            "type": "number",
            "minimum": 1
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64
          },
          "image": {
            "type": "string"
          }
        }
      },
      "Record": {
        "type": "object",
        "properties": {
          "recordId": {
            "type": "number"
          },
          "productId": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "unitPrice": {
            "type": "number"
          },
          "unitOfMeasure": {
            "type": "string"
          },
          "trackNumber": {
            "type": "string"
          }
        },
        "required": [
          "recordId",
          "productId",
          "createdAt",
          "unitPrice",
          "unitOfMeasure",
          "trackNumber"
        ]
      },
      "CreateRecordDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "number",
            "minimum": 1
          },
          "unitOfMeasure": {
            "type": "string"
          },
          "unitPrice": {
            "type": "number",
            "minimum": 1
          }
        },
        "required": [
          "productId",
          "unitOfMeasure",
          "unitPrice"
        ]
      },
      "UpdateRecordDto": {
        "type": "object",
        "properties": {
          "unitPrice": {
            "type": "number",
            "minimum": 1
          },
          "unitOfMeasure": {
            "type": "string"
          }
        }
      },
      "UploadResponse": {
        "type": "object",
        "properties": {
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path"
        ]
      },
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 50
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string",
            "minLength": 6
          },
          "taxId": {
            "type": "string",
            "minLength": 8,
            "maxLength": 8
          }
        },
        "required": [
          "name",
          "email",
          "password",
          "address",
          "phone",
          "taxId"
        ]
      },
      "AuthResponse": {
        "type": "object",
        "properties": {
          "jwt": {
            "type": "string"
          }
        },
        "required": [
          "jwt"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string",
            "minLength": 6,
            "maxLength": 50
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "FindMeResponse": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "taxId": {
            "type": "string"
          },
          "isAdmin": {
            "type": "boolean"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "name",
          "email",
          "address",
          "phone",
          "taxId",
          "isAdmin",
          "createdAt"
        ]
      }
    }
  }
}